ai/ai-server/requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.1
cycler==0.12.1
filelock==3.13.3
fonttools==4.51.0
fsspec==2024.3.1
idna==3.6
Jinja2==3.1.3
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.4
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
opencv-python==4.9.0.80
packaging==24.0
pandas==2.2.1
pika==1.3.2
pillow==10.3.0
psutil==5.9.8
py-cpuinfo==9.0.0
py-redis==1.1.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.3
requests==2.31.0
scipy==1.13.0
seaborn==0.13.2
six==1.16.0
sympy==1.12
thop==0.1.1.post2209072238
torch==2.2.2
torchvision==0.17.2
tqdm==4.66.2
typing_extensions==4.11.0
tzdata==2024.1
ultralytics==8.1.45
urllib3==2.2.1
